+++ mcan/CHANGELOG.md
@@ -3,7 +3,10 @@
 Tagging in git follows a pattern: `mcan/<version>`.
 
 ## [Unreleased]
-- Added support for fractional baudrates (#59)
+
+### Added
+- Add support for fractional baudrates (#59)
+- Add method to disable automatic retransmission (#60)
 
 ## [0.7.0] - 2025-04-23